Move useFeatures/withFeatures to separate files
