+++ src/index.js
@@ -2,9 +2,6 @@ export { configureFeature } from './configure-feature';
 export { FeatureToggles } from './feature-toggles';
 export { Feature } from './feature';
 export { withFeatureToggles } from './with-feature-toggles';
-export {
-  withFeatures,
-  useFeatures,
-  Consumer as FeaturesConsumer,
-  Provider as FeaturesProvider
-} from './context';
+export { FeatureTogglesContext } from './context';
+export { withFeatures } from './with-features';
+export { useFeatures } from './use-features';
