+++ src/use-features.js
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { FeatureTogglesContext } from './context';
+
+export const useFeatures = () => useContext(FeatureTogglesContext);
